\name{kddiscretize}
\alias{kddiscretize}

\title{kddiscretize Function}

\description{discretize a Time Series..}

\usage{
kddiscretize <- function (TSerie, Threshold,Labels)
}

\arguments{
\item{TSerie}{Time Series.}
\item{Threshold}{margins for chaging the label}
\item{Labels}{characters for describing the different levels of Threshold.}
}

\value{This function discretize a time series bering in mind the thersholds that can be selected.We can use labels for differing the levels.}
\author{
Francisco Javier Martinez de Pisón. \cr\email{francisco.martinez@dim.unirioja.es}\cr

Miguel Lodosa Ayala.                 \cr\email{miguelodosa@hotmail.com}\cr
}


\references{
}

\seealso{
}
\keyword{}

