\name{kdplotdiscretize}
\alias{kdplotdiscretize}

\title{kdplotdiscretize Function}

\description{Plot de original time series and the discretized Time Series..}

\usage{
kdplotdiscretize <- function (TSerieD,TSerie,Threshold,Labels)
}

\arguments{
\item{TSerieD}{Discretized time Series.}
\item{TSerie}{Original time Series.}
\item{Threshold}{margins for chaging the label}
\item{Labels}{characters for describing the different levels of Threshold.}
}

\value{This function plots the original time series and the discretized time series obteining from kddiscretize function.}
\author{
Francisco Javier Martinez de Pisón. \cr\email{francisco.martinez@dim.unirioja.es}\cr

Miguel Lodosa Ayala.                 \cr\email{miguelodosa@hotmail.com}\cr
}


\references{
}

\seealso{
}
\keyword{}

