\name{kdplotfilter}
\alias{kdplotfilter}

\title{Kdplotfilter Function}

\description{Function to draw the time series.}

\usage{
kdplotfilter <- function(TSerie,TSerieOrigin)
}

\arguments{
\item{TSerie}{Time Series.}
\item{TSerieOrigin}{Original time series.}
}

\value{This function plot in the same frame the original time series (without filtered) and the filtered time series (from kdfilter time series) so we can compare both of them.}
\author{
Francisco Javier Martinez de Pisón. \cr\email{francisco.martinez@dim.unirioja.es}\cr

Miguel Lodosa Ayala.                 \cr\email{miguelodosa@hotmail.com}\cr
}

\references{
}

\seealso{
}
\keyword{}

