\name{kdplotnumz}
\alias{kdplotnumz}

\title{kdplotnumz Function}

\description{Plot Number ZerosCrossing Curve vs Filter's Windows Width}

\usage{
kdplotnumz <- function(MAT, Positions=1:dim(MAT$MFilt)[1])
}

\arguments{
\item{MAT}{Matrix with Differents Filters(from kdmatfilter()).}
\item{Positions}{Plot filter placed in Positions.}
}

\value{Plot relative maximums and minimums of the time series that has been filtered by different windows width.}
\author{
Francisco Javier Martinez de Pisón. \cr\email{francisco.martinez@dim.unirioja.es}\cr

Miguel Lodosa Ayala.                 \cr\email{miguelodosa@hotmail.com}\cr
}

\references{
}

\seealso{
}
\keyword{}

