\name{kdplotpatt}
\alias{kdplotpatt}

\title{kdplotpatt Function}

\description{Plot Segments Extracted from kdextract()}

\usage{
kdplotpatt <- function (Pat, Ini=1, End=length(Pat$Serie))
}

\arguments{
\item{Pat}{Different type of segments found (FROM kdclusterpatt()).}
\item{Ini}{First point.}
\item{End}{Last point.}
}

\value{.}
\author{
Francisco Javier Martinez de Pisón. \cr\email{francisco.martinez@dim.unirioja.es}\cr

Miguel Lodosa Ayala.                 \cr\email{miguelodosa@hotmail.com}\cr
}

\references{
}

\seealso{
}
\keyword{}

