\name{kdplotzcross}
\alias{kdplotzcross}

\title{kdplotzcross Function}

\description{Plot Time Series and ZeroCrossings}

\usage{
kdplotzcross <- function (TSerie)
}

\arguments{
\item{TSerie}{Time series.}
}

\value{Plot relative maximums and minimums of the time series by simbols "+" (relative maximums and minimums) and "-" (the rest).}
\author{
Francisco Javier Martinez de Pisón. \cr\email{francisco.martinez@dim.unirioja.es}\cr

Miguel Lodosa Ayala.                 \cr\email{miguelodosa@hotmail.com}\cr
}

\references{
}

\seealso{
}
\keyword{}

